requirements.txt
langchain-aws
boto3
fastapi
pydantic
psycopg2-binary
uvicorn
python-dotenv
bcrypt
pydantic[email]
langchain-google-genai
langchain_nvidia_ai_endpoints
langchain
langchain-core
langchain-community
unstructured
opencv-python
langchain-chroma
markdown